examples/flows/chat/chat-with-pdf-maf/requirements.txt
agent-framework>=1.0.1
agent-framework-openai>=1.0.1
python-dotenv
PyPDF2
faiss-cpu
openai
jinja2
tiktoken
requests
